Separated the qt-centric classes from the model.
PoissonDiskSampling now has templated metaprogramming to deduct number
of dimensions and calculate the backgroundGrid accordingly.
The gameManager is not directly accessed from qml, instead there is a
gameManagerQMLBinding that acts as intermediary.
New ModelFacade that handles all the interactions between the model and
the outside world.
Update the cmake file to use CXX14